Move `Format` type under the `util` package
